Add tune "run" in order to run Herwig event generation on existing run files
The build phase for NLO processes can take a huge
amount of time. Doing this for each job as a separate
job is extremely inefficient and will lead to the job being
killed due to exceeding of the time limit. Instead this
step can be nicely parallelized on local clusters, making
this step much more efficient, so that for the processing
on the grid only the herwig.run file is needed. Therefore
the tune "run" skips herwig configuration and expects
a herwig.run file to be present in the current folder.

In addition add NLO dijet tunes using OpenLoops
or MadGraph as external loop providers.