Set long-lived particles to stable in Herwig
Set long-lived particles K0, Lambda, Sigma(0+-),
Xi(0-) and Omega to stable. Also always setting
corresponding antiparticle to stable. Definition of
stable particles is based on the definition of stable
particles in AliPythia8.

Due to the max. liftime setting of 10 mm heavy-flavour
particles are decayed by Herwig and are not set to
stable in case the HF mode is selected.