Adjust handling of packages for official and user builds
The behaviour of alienv with respect to modulefiles
is different for official and user builds. In case of user
builds by default alienv updates the module directory,
which is unwanted in particular when running multiple
jobs with aligenmc in parallel on a batch cluster.
Therefore in this case the option "--no-refresh" should
be specified.

In case of official packages from cvmfs a different
alienv is used, which does not support the option
"--no-refresh", leading to a failure of the execution.
Therefore the option must not be specified in case
of official pacakges.

The two modes can be distinguished by the package
name: Official packages split package and version
via "::", local builds via "/". It is expected that local
and cvmfs packages are not mixed.